реализую тест
