+++ tests/functional/put_v1_account_token/test_put_v1_account_token.py
@@ -1,15 +1,13 @@
 from json import loads
 from dm_api_account.apis.account_api import AccountApi
-from dm_api_account.apis.login_api import LoginApi
 from api_mailhog.apis.mailhog_api import MailhogApi
 
-def test_post_v1_account():
+def test_put_v1_account_token():
     # Регистрация пользователя
     account_api = AccountApi(host='http://5.63.153.31:5051')
-    login_api = LoginApi(host='http://5.63.153.31:5051')
     mailhog_api = MailhogApi(host='http://5.63.153.31:5025')
 
-    login = 's-test6'
+    login = 's-test7'
     password = '12345678'
     email = f'{login}@m.ru'
     json_data = {
@@ -19,15 +17,11 @@ def test_post_v1_account():
     }
 
     response = account_api.post_v1_account(json_data=json_data)
-    print(response.status_code)
-    print(response.text)
 
     assert response.status_code == 201, f"Пользователь не создан {response.json()}"
-    # Получить письма из почтового сервера
 
+    # Получить письма из почтового сервера
     response = mailhog_api.get_api_v2_messages()
-    print(response.status_code)
-    print(response.text)
     assert response.status_code == 200, "Письма не получены"
 
     # Получить активационный токен
@@ -37,34 +31,18 @@ def test_post_v1_account():
 
     # Активация пользователя
     response = account_api.put_v1_account_token(token=token)
-    print(response.status_code)
-    print(response.text)
 
     assert response.status_code == 200, "Пользователь не был активирован"
 
     # Авторизация
-    json_data = {
-        'login': login,
-        'password': password,
-        'rememberMe': True,
-    }
 
-    response = login_api.post_v1_account_login(json_data=json_data)
-    print(response.status_code)
-    print(response.text)
-    assert response.status_code == 200, "Пользователь не был авторизован"
-
-
-def get_activation_token_by_login(login, response):
+def get_activation_token_by_login(
+        login,
+        response
+        ):
     for item in response.json()['items']:
         user_data = loads(item['Content']['Body'])
         user_login = user_data['Login']
         if user_login == login:
             token = user_data['ConfirmationLinkUrl'].split('/')[-1]
     return token
-
-
-
-
-
-
